% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brgysub.R
\name{brgysub}
\alias{brgysub}
\title{brgysub Function
This function allows you subtitute barangay names on barangay columns.}
\usage{
brgysub(m, b)
}
\arguments{
\item{CBMS}{CORE HPQ VN05201701 main.csv}
}
\description{
brgysub Function
This function allows you subtitute barangay names on barangay columns.
}
\examples{
brgysub()

brgysub()
}
\keyword{cbms}
