% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{dedup Function
This function allows you get possible duplicates using jarowinkler algorithm on RecordLinkage Package}
\usage{
dedup(hpq, main)
}
\arguments{
\item{CBMS}{CORE HPQ VN05201701 main.csv}
}
\description{
dedup Function
This function allows you get possible duplicates using jarowinkler algorithm on RecordLinkage Package
}
\examples{
dedup()

dedup()
}
\keyword{cbms}
